+++ src/egg/sparkles.ts
@@ -1,9 +1,12 @@
 
 import { addScriptHook, W3TS_HOOK } from "w3ts";
+import { WISP_TYPE } from "../shared";
+import { WARD_TYPE } from "../wolves/ward";
 
 const onUnitCreated = (): boolean => {
 
 	const u = GetFilterUnit();
+	const uType = GetUnitTypeId( u );
 
 	if ( IsUnitType( u, UNIT_TYPE_STRUCTURE ) ) return false;
 
@@ -16,8 +19,12 @@ const onUnitCreated = (): boolean => {
 
 		AddSpecialEffectTarget( "Abilities\\Spells\\NightElf\\FaerieDragonInvis\\FaerieDragon_Invis.mdl", u, "origin" );
 
-	}/* else if ( playerName.indexOf( "hawkys" ) >= 0 )
-		AddSpecialEffectTarget( "Doodads\\Ruins\\Plants\\Ruins_Flower\\Ruins_Flower2.mdl", u, "head" );*/
+	} else if ( playerName.indexOf( "hawkys" ) >= 0 && uType !== WISP_TYPE && uType !== WARD_TYPE ) {
+
+		const e = AddSpecialEffectTarget( "Doodads\\Underground\\Plants\\ShroomsBlue\\ShroomsBlue.mdl", u, "head" );
+		BlzSetSpecialEffectScale( e, 0.1 );
+
+	}
 
 	return false;